Refactor VariableBox into sub components
