+++ libs/pxweb2-ui/src/lib/components/VariableBox/VariableBox.module.scss
@@ -1,27 +1,3 @@
-@use '../../../../style-dictionary/dist/scss/fixed-variables.scss' as fixed;
-
-/* ANIMATIONS */
-@keyframes SlideDown {
-  from {
-    transform: translateY(-100%);
-  }
-  to {
-    transform: translateY(0);
-  }
-}
-
-@keyframes slideUp {
-  from {
-    position: sticky;
-    transform: translateY(0);
-  }
-  to {
-    transform: translateY(-100%);
-    position: inherit;
-  }
-}
-
-/* STYLES */
 .variablebox {
   /* Layout */
   display: flex;
@@ -46,122 +22,4 @@
     border-radius: var(--px-border-radius-medium);
     background: var(--px-color-surface-action-subtle-hover);
   }
-}
-
-.variablebox-header {
-  /* Layout */
-  display: flex;
-  align-items: flex-start;
-  align-self: stretch;
-  padding: fixed.$spacing-3 fixed.$spacing-3 fixed.$spacing-5 fixed.$spacing-5;
-  gap: 1rem;
-
-  /* Style */
-  background: var(--px-color-surface-default);
-  border-radius: var(--px-border-radius-large);
-  outline: none; // Remove focus outline, handled with focus marking
-}
-
-.variablebox-header-isopen {
-  border-radius: var(--px-border-radius-large) var(--px-border-radius-large) 0 0;
-}
-
-.header-title-and-tag {
-  /* Layout */
-  display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  align-self: stretch;
-  flex: 1 0 0;
-  padding-top: 0.375rem;
-  gap: fixed.$spacing-2;
-}
-
-.header-title {
-  /* Layout */
-  display: flex;
-  align-items: flex-start;
-  align-self: stretch;
-  padding-top: 0.125rem;
-}
-
-.header-tags {
-  /* Layout */
-  display: flex;
-  align-items: flex-start;
-  gap: fixed.$spacing-2;
-}
-
-.header-icon {
-  /* Layout */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  width: 2.75rem;
-  height: 2.75rem;
-}
-
-.variablebox-content {
-  /* Layout */
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  align-self: stretch;
-  padding: fixed.$spacing-2 fixed.$spacing-5 fixed.$spacing-5 fixed.$spacing-5;
-  gap: fixed.$spacing-3;
-
-  /* Style */
-  background: var(--px-color-surface-default);
-  border-radius: 0 0 var(--px-border-radius-large) var(--px-border-radius-large);
-}
-
-.variablebox-content-main {
-  display: flex;
-  flex-direction: column;
-  align-self: stretch;
-
-  /* Style */
-  border-radius: var(--px-border-radius-medium);
-  border: 2px solid var(--px-color-border-subtle);
-}
-
-.variablebox-content-select {
-  align-self: stretch;
-}
-
-.variablebox-content-values-list {
-  /* Layout */
-  display: flex;
-  flex-direction: column;
-  align-self: stretch;
-}
-
-.variablebox-content-values-list-scroll {
-  position: relative;
-  max-height: 380px;
-  overflow-y: auto;
-}
-
-.variablebox-content-values-list-scroll-up {
-  position: sticky;
-  top: 0;
-  z-index: 1;
-  animation: SlideDown 0.2s;
-}
-
-.variablebox-content-values-list-scroll-down {
-  //position: block;
-  top: 0;
-  z-index: 1;
-  animation: slideUp 0.2s;
-}
-
-.variablebox-content-metadata-link {
-  /* Layout */
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  align-self: stretch;
-  gap: fixed.$spacing-1;
 }
