+++ libs/pxweb2-ui/src/lib/components/VariableBox/VariableBoxContent/VariableBoxContent.module.scss
@@ -1,0 +1,86 @@
+@use '../../../../../style-dictionary/dist/scss/fixed-variables.scss' as fixed;
+
+/* ANIMATIONS */
+@keyframes SlideDown {
+  from {
+    transform: translateY(-100%);
+  }
+  to {
+    transform: translateY(0);
+  }
+}
+
+@keyframes slideUp {
+  from {
+    position: sticky;
+    transform: translateY(0);
+  }
+  to {
+    transform: translateY(-100%);
+    position: inherit;
+  }
+}
+
+/* STYLES */
+.variablebox-content {
+  /* Layout */
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  align-self: stretch;
+  padding: fixed.$spacing-2 fixed.$spacing-5 fixed.$spacing-5 fixed.$spacing-5;
+  gap: fixed.$spacing-3;
+
+  /* Style */
+  background: var(--px-color-surface-default);
+  border-radius: 0 0 var(--px-border-radius-large) var(--px-border-radius-large);
+}
+
+.variablebox-content-main {
+  display: flex;
+  flex-direction: column;
+  align-self: stretch;
+
+  /* Style */
+  border-radius: var(--px-border-radius-medium);
+  border: 2px solid var(--px-color-border-subtle);
+}
+
+.variablebox-content-select {
+  align-self: stretch;
+}
+
+.variablebox-content-values-list {
+  /* Layout */
+  display: flex;
+  flex-direction: column;
+  align-self: stretch;
+}
+
+.variablebox-content-values-list-scroll {
+  position: relative;
+  max-height: 380px;
+  overflow-y: auto;
+}
+
+.variablebox-content-values-list-scroll-up {
+  position: sticky;
+  top: 0;
+  z-index: 1;
+  animation: SlideDown 0.2s;
+}
+
+.variablebox-content-values-list-scroll-down {
+  top: 0;
+  z-index: 1;
+  animation: slideUp 0.2s;
+}
+
+.variablebox-content-metadata-link {
+  /* Layout */
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  align-self: stretch;
+  gap: fixed.$spacing-1;
+}
