+++ libs/pxweb2-ui/src/lib/components/VariableBox/VariableBoxContent/VariableBoxContent.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import VariableBoxContent from './VariableBoxContent';
+
+describe('VariableBoxContent', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<VariableBoxContent />);
+    expect(baseElement).toBeTruthy();
+  });
+});
