+++ libs/pxweb2-ui/src/lib/components/VariableBox/VariableBoxHeader/VariableBoxHeader.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import VariableBoxHeader from './VariableBoxHeader';
+
+describe('VariableBoxHeader', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<VariableBoxHeader />);
+    expect(baseElement).toBeTruthy();
+  });
+});
